requirements.txt
llama-index
together
python-dotenv